[#62] refactor: 채팅방 목록 조회 메소드 ChatroomListService로 분리
